+++ src/RcsbFvWeb/RcsbFvFactories/RcsbFvBlockFactory/AlignmentBlockFactory.ts
@@ -7,14 +7,14 @@ import {BlockFactoryInterface} from "./BlockFactoryInterface";
 import {RcsbFvRowConfigInterface} from "@rcsb/rcsb-saguaro";
 import {TrackFactoryInterface} from "../RcsbFvTrackFactory/TrackFactoryInterface";
 
-export class AlignmentBlockFactory implements BlockFactoryInterface<[AlignmentRequestContextType, AlignmentResponse],[AlignmentRequestContextType, TargetAlignment]> {
+export class AlignmentBlockFactory implements BlockFactoryInterface<[AlignmentRequestContextType, AlignmentResponse],[AlignmentRequestContextType, TargetAlignment, AlignmentResponse]> {
 
     readonly trackFactory: TrackFactoryInterface<[AlignmentRequestContextType, TargetAlignment]>;
-    readonly trackConfigModifier?: (alignmentContext: AlignmentRequestContextType, targetAlignment: TargetAlignment) => Promise<Partial<RcsbFvRowConfigInterface>>;
+    readonly trackConfigModifier?: (alignmentContext: AlignmentRequestContextType, targetAlignment: TargetAlignment, alignmentResponse: AlignmentResponse) => Promise<Partial<RcsbFvRowConfigInterface>>;
 
     constructor(
         alignmentTrackFactory: TrackFactoryInterface<[AlignmentRequestContextType, TargetAlignment]>,
-        trackModifier?: (alignmentContext: AlignmentRequestContextType, targetAlignment: TargetAlignment) => Promise<Partial<RcsbFvRowConfigInterface>>
+        trackModifier?: (alignmentContext: AlignmentRequestContextType, targetAlignment: TargetAlignment, alignmentResponse: AlignmentResponse) => Promise<Partial<RcsbFvRowConfigInterface>>
     ) {
         this.trackFactory = alignmentTrackFactory;
         this.trackConfigModifier = trackModifier;
@@ -31,7 +31,7 @@ export class AlignmentBlockFactory implements BlockFactoryInterface<[AlignmentRe
                 return;
            return {
                ... await this.trackFactory.getTrack(alignmentRequestContext, alignment),
-               ... await this.trackConfigModifier?.(alignmentRequestContext,alignment)
+               ... await this.trackConfigModifier?.(alignmentRequestContext,alignment, alignmentData)
            };
         }))).filter((x): x is RcsbFvRowConfigInterface=>x!=undefined);
     }